104-fibonacci.c

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,7 +7,7 @@
  */
 int main(void)
 {
-	unsigned long int fib[98];
+	unsigned long long int fib[98];
 	int count;
 
 	fib[0] = 1;
